Removes reference to datatable statics
The duplicate references generated to these statics in the page body cause global datatable defaults to be lost.  This is particularly noticeable in the bootstrap site as we normalize the way we include the statics in the base template.

axis/master has been updated so that its base template includes these resources on every page, removing the pileup on the global settings battle.